requirements.txt
# Core dependencies
PyQt6>=6.4.0
PyOpenGL>=3.1.6
numpy>=1.24.0
trimesh>=3.9.0

# Additional dependencies
scipy>=1.10.0
networkx>=2.8.0
Pillow>=9.3.0
matplotlib>=3.6.0

# Development dependencies
pytest>=7.3.0
black>=22.3.0
flake8>=6.0.0